Further work on menus and beginning of snake.py refactoring
Originally I had not intended on making a menu or any features beyond
the plain game of snake, but I've decided to do so and the implementation
I have of snake is messy and needs to be moved around. It will become
a class and have global variables changed to fields and just generally
cleaned up.