Created a textbox class since online ones sucked
Further documentation will follow but it's late and I don't want
to write a full message right now.  It accepts key presses for
alphanumeric characters and most symbols.  Shift works for
capitalization.  Currently RETURN should be returning the text
contents, but that seems to not be working.  I'll be adding a bit
more functionality and customization options in the near future.