FINALLY finished high score features(kinda)
I just realized as I a typing this that there is not yet a back
button on the high scores page, but I'll fix that.  Anyway, the
scores are now stored in a list of tuples rather than two seperate
lists. This makes iterating much easier and passing things around
more straightforward as well.  The scores sort properly and will
display based on the scoresOnScreen value in ScoreDisplay.py.