Moveable cursor working. Draws incorrectly
The cursor draws in each position it has been in starting from the
right of the inputString.  I'm trying to track down the problem.
Also, occasionally the cursor doesn't show at all even though the
string can be editted as normal.  Not sure what exactly is causing
this but I'm looking around.